replace utils
