+++ mlt/__init__.py
@@ -2,3 +2,4 @@ from mlt import stats
 from mlt import file
 from mlt import models
 from mlt import preprocessing
+from mlt import utils
