+++ mlt/preprocessing/FeatureTransform.py
@@ -1,4 +1,3 @@
-import sklearn.preprocessing
 import pandas as pd
