+++ mlt/utils/weight.py
@@ -1,0 +1,6 @@
+import numpy as np
+
+
+def init_zeros(n):
+    w = np.zeros(n, dtype=float)
+    return w
